Enhancement: Extract IsEqualTest
